move Defer to new file to break circular dependency
