+++ src/util/promise.js
@@ -210,56 +210,6 @@ async function retry({ attempts, delay_ms, func, error_logger }) {
 /////////////////////////////////////
 
 /**
- * @deprecated LEGACY PROMISE UTILS - DEPRECATED IN FAVOR OF ASYNC-AWAIT
- */
-class Defer {
-
-    constructor() {
-        this.isPending = true;
-        this.isResolved = false;
-        this.isRejected = false;
-        this.promise = new Promise((resolve, reject) => {
-            this._promise_resolve = resolve;
-            this._promise_reject = reject;
-        });
-        Object.seal(this);
-    }
-
-    // setting resolve and reject to assert that the current code assumes 
-    // the Promise ctor is calling the callback synchronously and not deferring it,
-    // otherwise we might have weird cases that we miss the caller's resolve/reject
-    // events, so we throw to assert 
-
-    /**
-     * @param {any} [res]
-     * @returns {void}
-     */
-    resolve(res) {
-        if (!this.isPending) {
-            return;
-        }
-        this.isPending = false;
-        this.isResolved = true;
-        Object.freeze(this);
-        this._promise_resolve(res);
-    }
-
-    /**
-     * @param {Error} err
-     * @returns {void}
-     */
-    reject(err) {
-        if (!this.isPending) {
-            return;
-        }
-        this.isPending = false;
-        this.isRejected = true;
-        Object.freeze(this);
-        this._promise_reject(err);
-    }
-}
-
-/**
  * Callback is a template typedef to help propagate types correctly
  * when using nodejs callback functions.
  * @template T
@@ -366,5 +316,4 @@ exports.fromCallback = fromCallback; // 44 occurrences
 exports.fcall = fcall; // 59 occurrences
 exports.ninvoke = ninvoke; // 30 occurrences
 exports.wait_until = wait_until; // 20 occurrences
-exports.Defer = Defer; // 13 occurrences
 exports.pwhile = pwhile; // 15 occurrences
